#162 merge api activity calls
