--- get_stock_data_orchestrator/function.json
@@ -1,10 +1,0 @@
-{
-  "scriptFile": "__init__.py",
-  "bindings": [
-    {
-      "name": "context",
-      "type": "orchestrationTrigger",
-      "direction": "in"
-    }
-  ]
-}
